refactor: split string/char primitives into separate files
Extract string and character primitives from the monolithic prim_all.go
into prim_strings.go and prim_characters.go. The generic variadicCompare
helper and record/promise code remain in prim_all.go.

Closes #27
