+++ go/extensions/all/prim_all.go
@@ -12,19 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Additional R7RS primitives for the all extension:
-// Records, promises, string operations, character operations
+// Records, promises, and generic variadic comparison helper
 
 package all
 
 import (
 	"context"
 	"errors"
-	"strings"
-	"unicode"
-
-	"golang.org/x/text/cases"
-	"golang.org/x/text/language"
 
 	"github.com/aalpar/wile/go/environment"
 	"github.com/aalpar/wile/go/machine"
@@ -436,598 +430,5 @@ func variadicCompare[T any, V values.Value](
 	}
 
 	mc.SetValue(values.TrueValue)
-	return nil
-}
-
-// =============================================================================
-// String Primitives
-// =============================================================================
-
-// stringCompareVariadic is a helper for variadic string comparison primitives.
-func stringCompareVariadic(mc *machine.MachineContext, name string, cmp func(a, b string) bool) error {
-	return variadicCompare(mc, name,
-		func(v values.Value) (*values.String, bool) {
-			s, ok := v.(*values.String)
-			return s, ok
-		},
-		func(s *values.String) string {
-			return s.Value
-		},
-		cmp,
-		values.ErrNotAString,
-		"a string")
-}
-
-// PrimStringCopyTo implements the string-copy! primitive.
-// R7RS §6.7: (string-copy! to at from [start [end]])
-func PrimStringCopyTo(_ context.Context, mc *machine.MachineContext) error {
-	toArg := mc.Arg(0)
-	rest := mc.Arg(1)
-
-	to, ok := toArg.(*values.String)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAString, "string-copy!: expected a string for 'to' but got %T", toArg)
-	}
-
-	// Parse variadic arguments: at from [start [end]]
-	var args []values.Value
-	current := rest
-	for current != values.EmptyList {
-		tuple, ok := current.(values.Tuple)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotAList, "string-copy!: improper argument list")
-		}
-		args = append(args, tuple.Car())
-		current = tuple.Cdr()
-	}
-
-	if len(args) < 2 {
-		return values.NewForeignError("string-copy!: expected at least 3 arguments")
-	}
-
-	atVal, ok := args[0].(*values.Integer)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotANumber, "string-copy!: expected an integer for 'at' but got %T", args[0])
-	}
-	at := int(atVal.Value)
-
-	from, ok := args[1].(*values.String)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAString, "string-copy!: expected a string for 'from' but got %T", args[1])
-	}
-
-	fromLen := from.Len()
-	start := 0
-	end := fromLen
-
-	if len(args) >= 3 {
-		startVal, ok := args[2].(*values.Integer)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotANumber, "string-copy!: expected an integer for start but got %T", args[2])
-		}
-		start = int(startVal.Value)
-	}
-
-	if len(args) >= 4 {
-		endVal, ok := args[3].(*values.Integer)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotANumber, "string-copy!: expected an integer for end but got %T", args[3])
-		}
-		end = int(endVal.Value)
-	}
-
-	// Validate indices
-	if start < 0 || end > fromLen || start > end {
-		return values.NewForeignError("string-copy!: invalid source indices")
-	}
-
-	toLen := to.Len()
-	copyLen := end - start
-	if at < 0 || at+copyLen > toLen {
-		return values.NewForeignError("string-copy!: destination index out of bounds")
-	}
-
-	// Perform the copy
-	toRunes := to.Runes()
-	fromRunes := from.Runes()
-	copy(toRunes[at:], fromRunes[start:end])
-	to.SetValue(string(toRunes))
-
-	mc.SetValue(values.Void)
-	return nil
-}
-
-// PrimStringFill implements the string-fill! primitive.
-// R7RS §6.7: (string-fill! string fill [start [end]])
-func PrimStringFill(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	rest := mc.Arg(1)
-
-	s, ok := o.(*values.String)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAString, "string-fill!: expected a string but got %T", o)
-	}
-
-	// Parse variadic arguments: fill [start [end]]
-	var args []values.Value
-	current := rest
-	for current != values.EmptyList {
-		tuple, ok := current.(values.Tuple)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotAList, "string-fill!: improper argument list")
-		}
-		args = append(args, tuple.Car())
-		current = tuple.Cdr()
-	}
-
-	if len(args) < 1 {
-		return values.NewForeignError("string-fill!: expected at least 2 arguments")
-	}
-
-	char, ok := args[0].(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "string-fill!: expected a character but got %T", args[0])
-	}
-
-	length := s.Len()
-	start := 0
-	end := length
-
-	if len(args) >= 2 {
-		startVal, ok := args[1].(*values.Integer)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotANumber, "string-fill!: expected an integer for start but got %T", args[1])
-		}
-		start = int(startVal.Value)
-	}
-
-	if len(args) >= 3 {
-		endVal, ok := args[2].(*values.Integer)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotANumber, "string-fill!: expected an integer for end but got %T", args[2])
-		}
-		end = int(endVal.Value)
-	}
-
-	if start < 0 || end > length || start > end {
-		return values.NewForeignError("string-fill!: invalid indices")
-	}
-
-	s.Fill(char.Value, start, end)
-	mc.SetValue(values.Void)
-	return nil
-}
-
-// PrimStringMap implements the string-map primitive.
-// R7RS §6.7: (string-map proc string1 string2 ...)
-func PrimStringMap(_ context.Context, mc *machine.MachineContext) error {
-	proc := mc.Arg(0)
-	stringsVal := mc.Arg(1)
-
-	mcls, ok := proc.(*machine.MachineClosure)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAProcedure, "string-map: expected a procedure but got %T", proc)
-	}
-
-	if values.IsEmptyList(stringsVal) {
-		return values.WrapForeignErrorf(values.ErrWrongNumberOfArguments, "string-map: expected at least one string")
-	}
-
-	// Collect all strings into a slice
-	var strs []*values.String
-	current := stringsVal
-	for !values.IsEmptyList(current) {
-		tuple, ok := current.(values.Tuple)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotAList, "string-map: improper argument list")
-		}
-		s, ok := tuple.Car().(*values.String)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotAString, "string-map: expected a string but got %T", tuple.Car())
-		}
-		strs = append(strs, s)
-		current = tuple.Cdr()
-	}
-
-	if len(strs) == 0 {
-		return values.WrapForeignErrorf(values.ErrWrongNumberOfArguments, "string-map: expected at least one string")
-	}
-
-	// Convert all strings to rune slices and find minimum length
-	runeSlices := make([][]rune, len(strs))
-	minLen := -1
-	for i, s := range strs {
-		runeSlices[i] = s.Runes()
-		if minLen < 0 || len(runeSlices[i]) < minLen {
-			minLen = len(runeSlices[i])
-		}
-	}
-
-	// Apply proc to each position
-	result := make([]rune, minLen)
-	sub := mc.NewSubContext()
-
-	for i := 0; i < minLen; i++ {
-		// Collect one character from each string
-		args := make(values.Vector, len(strs))
-		for j := range strs {
-			args[j] = values.NewCharacter(runeSlices[j][i])
-		}
-
-		// Apply proc to collected arguments
-		_, err := sub.Apply(mcls, args...)
-		if err != nil {
-			return err
-		}
-		err = sub.Run()
-		if err != nil {
-			var escapeErr *machine.ErrContinuationEscape
-			if errors.As(err, &escapeErr) {
-				return err
-			}
-			if !errors.Is(err, machine.ErrMachineHalt) {
-				return err
-			}
-		}
-
-		// Get the result character
-		resultVal := sub.GetValue()
-		char, ok := resultVal.(*values.Character)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotACharacter, "string-map: procedure must return a character but got %T", resultVal)
-		}
-		result[i] = char.Value
-	}
-
-	mc.SetValue(values.NewString(string(result)))
-	return nil
-}
-
-// PrimStringForEach implements the string-for-each primitive.
-// R7RS §6.7: (string-for-each proc string1 string2 ...)
-func PrimStringForEach(_ context.Context, mc *machine.MachineContext) error {
-	proc := mc.Arg(0)
-	stringsVal := mc.Arg(1)
-
-	mcls, ok := proc.(*machine.MachineClosure)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAProcedure, "string-for-each: expected a procedure but got %T", proc)
-	}
-
-	if values.IsEmptyList(stringsVal) {
-		return values.WrapForeignErrorf(values.ErrWrongNumberOfArguments, "string-for-each: expected at least one string")
-	}
-
-	// Collect all strings into a slice
-	var strs []*values.String
-	current := stringsVal
-	for !values.IsEmptyList(current) {
-		tuple, ok := current.(values.Tuple)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotAList, "string-for-each: improper argument list")
-		}
-		s, ok := tuple.Car().(*values.String)
-		if !ok {
-			return values.WrapForeignErrorf(values.ErrNotAString, "string-for-each: expected a string but got %T", tuple.Car())
-		}
-		strs = append(strs, s)
-		current = tuple.Cdr()
-	}
-
-	if len(strs) == 0 {
-		return values.WrapForeignErrorf(values.ErrWrongNumberOfArguments, "string-for-each: expected at least one string")
-	}
-
-	// Convert all strings to rune slices and find minimum length
-	runeSlices := make([][]rune, len(strs))
-	minLen := -1
-	for i, s := range strs {
-		runeSlices[i] = s.Runes()
-		if minLen < 0 || len(runeSlices[i]) < minLen {
-			minLen = len(runeSlices[i])
-		}
-	}
-
-	// Apply proc to each position
-	sub := mc.NewSubContext()
-
-	for i := 0; i < minLen; i++ {
-		// Collect one character from each string
-		args := make(values.Vector, len(strs))
-		for j := range strs {
-			args[j] = values.NewCharacter(runeSlices[j][i])
-		}
-
-		// Apply proc to collected arguments
-		_, err := sub.Apply(mcls, args...)
-		if err != nil {
-			return err
-		}
-		err = sub.Run()
-		if err != nil {
-			var escapeErr *machine.ErrContinuationEscape
-			if errors.As(err, &escapeErr) {
-				return err
-			}
-			if !errors.Is(err, machine.ErrMachineHalt) {
-				return err
-			}
-		}
-	}
-
-	mc.SetValue(values.Void)
-	return nil
-}
-
-// PrimStringCiEqVariadic implements the variadic string-ci=? primitive.
-func PrimStringCiEqVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return stringCompareVariadic(mc, "string-ci=?", strings.EqualFold)
-}
-
-// PrimStringCiLtVariadic implements the variadic string-ci<? primitive.
-func PrimStringCiLtVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return stringCompareVariadic(mc, "string-ci<?", func(a, b string) bool {
-		return strings.ToLower(a) < strings.ToLower(b)
-	})
-}
-
-// PrimStringCiGtVariadic implements the variadic string-ci>? primitive.
-func PrimStringCiGtVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return stringCompareVariadic(mc, "string-ci>?", func(a, b string) bool {
-		return strings.ToLower(a) > strings.ToLower(b)
-	})
-}
-
-// PrimStringCiLeVariadic implements the variadic string-ci<=? primitive.
-func PrimStringCiLeVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return stringCompareVariadic(mc, "string-ci<=?", func(a, b string) bool {
-		return strings.ToLower(a) <= strings.ToLower(b)
-	})
-}
-
-// PrimStringCiGeVariadic implements the variadic string-ci>=? primitive.
-func PrimStringCiGeVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return stringCompareVariadic(mc, "string-ci>=?", func(a, b string) bool {
-		return strings.ToLower(a) >= strings.ToLower(b)
-	})
-}
-
-// PrimStringUpcase implements the string-upcase primitive.
-// R7RS §6.7: Returns a string whose characters are the uppercase versions of the characters in string.
-// Uses Unicode full case mapping which can expand characters (e.g., ß → SS).
-func PrimStringUpcase(_ context.Context, mc *machine.MachineContext) error {
-	s := mc.Arg(0)
-	str, ok := s.(*values.String)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAString, "string-upcase: expected a string but got %T", s)
-	}
-	// Use Unicode full case mapping (language-independent)
-	caser := cases.Upper(language.Und)
-	result := caser.String(str.Value)
-	mc.SetValue(values.NewString(result))
-	return nil
-}
-
-// PrimStringDowncase implements the string-downcase primitive.
-// R7RS §6.7: Returns a string whose characters are the lowercase versions of the characters in string.
-// Uses Unicode full case mapping which can expand characters.
-func PrimStringDowncase(_ context.Context, mc *machine.MachineContext) error {
-	s := mc.Arg(0)
-	str, ok := s.(*values.String)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAString, "string-downcase: expected a string but got %T", s)
-	}
-	// Use Unicode full case mapping (language-independent)
-	caser := cases.Lower(language.Und)
-	result := caser.String(str.Value)
-	mc.SetValue(values.NewString(result))
-	return nil
-}
-
-// PrimStringFoldcase implements the string-foldcase primitive.
-// R7RS §6.7: Returns a string whose characters are the case-folded versions of the characters in string.
-// Uses Unicode full case folding which can expand characters (e.g., ß → ss).
-func PrimStringFoldcase(_ context.Context, mc *machine.MachineContext) error {
-	s := mc.Arg(0)
-	str, ok := s.(*values.String)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotAString, "string-foldcase: expected a string but got %T", s)
-	}
-	// Use Unicode full case folding
-	caser := cases.Fold()
-	result := caser.String(str.Value)
-	mc.SetValue(values.NewString(result))
-	return nil
-}
-
-// =============================================================================
-// Character Primitives
-// =============================================================================
-
-// charCompareVariadic is a helper for variadic character comparison primitives.
-func charCompareVariadic(mc *machine.MachineContext, name string, cmp func(a, b rune) bool) error {
-	return variadicCompare(mc, name,
-		func(v values.Value) (*values.Character, bool) {
-			c, ok := v.(*values.Character)
-			return c, ok
-		},
-		func(c *values.Character) rune {
-			return c.Value
-		},
-		cmp,
-		values.ErrNotACharacter,
-		"a character")
-}
-
-// PrimCharCiEqVariadic implements the variadic char-ci=? primitive.
-func PrimCharCiEqVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return charCompareVariadic(mc, "char-ci=?", func(a, b rune) bool {
-		return unicode.ToLower(a) == unicode.ToLower(b)
-	})
-}
-
-// PrimCharCiLtVariadic implements the variadic char-ci<? primitive.
-func PrimCharCiLtVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return charCompareVariadic(mc, "char-ci<?", func(a, b rune) bool {
-		return unicode.ToLower(a) < unicode.ToLower(b)
-	})
-}
-
-// PrimCharCiGtVariadic implements the variadic char-ci>? primitive.
-func PrimCharCiGtVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return charCompareVariadic(mc, "char-ci>?", func(a, b rune) bool {
-		return unicode.ToLower(a) > unicode.ToLower(b)
-	})
-}
-
-// PrimCharCiLeVariadic implements the variadic char-ci<=? primitive.
-func PrimCharCiLeVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return charCompareVariadic(mc, "char-ci<=?", func(a, b rune) bool {
-		return unicode.ToLower(a) <= unicode.ToLower(b)
-	})
-}
-
-// PrimCharCiGeVariadic implements the variadic char-ci>=? primitive.
-func PrimCharCiGeVariadic(_ context.Context, mc *machine.MachineContext) error {
-	return charCompareVariadic(mc, "char-ci>=?", func(a, b rune) bool {
-		return unicode.ToLower(a) >= unicode.ToLower(b)
-	})
-}
-
-// PrimCharAlphabeticQ tests if a character is alphabetic.
-func PrimCharAlphabeticQ(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-alphabetic?: expected a character but got %T", o)
-	}
-	mc.SetValue(utils.BoolToBoolean(unicode.IsLetter(ch.Value)))
-	return nil
-}
-
-// PrimCharNumericQ tests if a character is numeric.
-func PrimCharNumericQ(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-numeric?: expected a character but got %T", o)
-	}
-	mc.SetValue(utils.BoolToBoolean(unicode.IsDigit(ch.Value)))
-	return nil
-}
-
-// PrimCharWhitespaceQ tests if a character is whitespace.
-func PrimCharWhitespaceQ(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-whitespace?: expected a character but got %T", o)
-	}
-	mc.SetValue(utils.BoolToBoolean(unicode.IsSpace(ch.Value)))
-	return nil
-}
-
-// PrimCharUpperCaseQ tests if a character is uppercase.
-func PrimCharUpperCaseQ(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-upper-case?: expected a character but got %T", o)
-	}
-	mc.SetValue(utils.BoolToBoolean(unicode.IsUpper(ch.Value)))
-	return nil
-}
-
-// PrimCharLowerCaseQ tests if a character is lowercase.
-func PrimCharLowerCaseQ(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-lower-case?: expected a character but got %T", o)
-	}
-	mc.SetValue(utils.BoolToBoolean(unicode.IsLower(ch.Value)))
-	return nil
-}
-
-// PrimCharUpcase returns the uppercase version of a character.
-func PrimCharUpcase(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-upcase: expected a character but got %T", o)
-	}
-	mc.SetValue(values.NewCharacter(unicode.ToUpper(ch.Value)))
-	return nil
-}
-
-// PrimCharDowncase returns the lowercase version of a character.
-func PrimCharDowncase(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-downcase: expected a character but got %T", o)
-	}
-	mc.SetValue(values.NewCharacter(unicode.ToLower(ch.Value)))
-	return nil
-}
-
-// PrimCharFoldcase returns the case-folded version of a character.
-// R7RS §6.6: Returns the simple Unicode case-folded version of the character.
-// Simple case folding maps each character to exactly one character.
-func PrimCharFoldcase(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "char-foldcase: expected a character but got %T", o)
-	}
-	mc.SetValue(values.NewCharacter(simpleCaseFold(ch.Value)))
-	return nil
-}
-
-// simpleCaseFold performs Unicode simple case folding on a rune.
-// Simple case folding maps each character to exactly one character.
-// This is used for case-insensitive matching as specified by R7RS.
-//
-// For most characters, simple case folding is equivalent to lowercase.
-// Special cases are handled based on Unicode CaseFolding.txt:
-//   - Capital sharp S (ẞ U+1E9E) folds to lowercase sharp s (ß U+00DF)
-//   - Most other characters just use ToLower
-func simpleCaseFold(r rune) rune {
-	// Handle special cases from Unicode CaseFolding.txt
-	switch r {
-	case 'ẞ': // U+1E9E LATIN CAPITAL LETTER SHARP S
-		return 'ß' // U+00DF LATIN SMALL LETTER SHARP S
-	case 'K': // U+212A KELVIN SIGN (if we get that far)
-		// Actually this is regular K, ignore
-	}
-	// For most characters, simple case folding equals lowercase
-	return unicode.ToLower(r)
-}
-
-// PrimDigitValue implements the (digit-value) primitive.
-// R7RS §6.6: Returns the numeric value (0-9) of a character that is a decimal digit
-// according to Unicode, or #f if it is not a decimal digit.
-func PrimDigitValue(_ context.Context, mc *machine.MachineContext) error {
-	o := mc.Arg(0)
-	ch, ok := o.(*values.Character)
-	if !ok {
-		return values.WrapForeignErrorf(values.ErrNotACharacter, "digit-value: expected a character but got %T", o)
-	}
-	// Check if it's a Unicode decimal digit (Nd category)
-	// Unicode decimal digits have values 0-9 within their respective scripts
-	if unicode.IsDigit(ch.Value) {
-		// Get the digit value by finding the base '0' for this script
-		// Unicode organizes decimal digits in blocks of 10: 0, 1, 2, ..., 9
-		base := ch.Value
-		for unicode.IsDigit(base - 1) {
-			base--
-		}
-		digit := int64(ch.Value - base)
-		if digit >= 0 && digit <= 9 {
-			mc.SetValue(values.NewInteger(digit))
-			return nil
-		}
-	}
-	mc.SetValue(values.FalseValue)
 	return nil
 }
